requirements.txt
requests>=2.31.0
pandas>=2.2.0
python-dotenv>=1.0.0
schedule>=1.2.0
anthropic>=0.7.8
tweepy>=4.14.0
textblob>=0.17.1
numpy>=1.26.0
streamlit>=1.28.0
plotly>=5.17.0
psutil>=5.9.0